fix: url pattern change
