+++ srcs/backend/srcs/mysite/follows/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path, include
+from rest_framework.routers import DefaultRouter
+from . import views
+
+urlpatterns = [
+    path('me', views.UserFollowView.as_view())
+]
