+++ srcs/backend/srcs/mysite/users/urls.py
@@ -9,5 +9,4 @@ router.register(r'\w+', views.UserViewSet, basename='user-detail')
 
 urlpatterns = [
     path('', include(router.urls)),
-    path('auth/follows', views.UserFollowView.as_view())
 ]
